Move layout to its own module
